gotwaterbot/requirements.txt
python-telegram-bot>=20.0
requests==2.31.0
python-dotenv==1.0.0
Pillow
overpy==0.6
geopy==2.4.1
